Conflicts:
	railo-java/railo-core/src/railo/runtime/Info.ini
	railo-java/railo-core/src/railo/runtime/functions/international/LSParseDateTime.java
	railo-java/railo-core/src/railo/runtime/interpreter/CFMLExpressionInterpreter.java
	railo-java/railo-core/src/railo/runtime/net/rpc/AxisCaster.java
	railo-java/railo-core/src/resource/context/railo-context.ra

+++ railo-java/railo-core/src/railo/runtime/tag/StoredProc.java
@@ -612,8 +612,8 @@ public class StoredProc extends BodyTagTryCatchFinallySupport {
 	/**
 	 * @param timeout the timeout to set
 	 */
-	public void setTimeout(int timeout) {
-		this.timeout = timeout;
+	public void setTimeout(double timeout) {
+		this.timeout = (int) timeout;
 	}
